Allow nil params and add ping to RequestRegistry (#25)
JSON-RPC allows params to be omitted:

> A Structured value that holds the parameter values to be used during the invocation of the method. **This member MAY be omitted.**

Previously, parsing was more strict and would fail if the member was omitted, but this change makes it optional.

In addition, the `ping` request has now been added to the `RequestRegistry` so an incoming ping from the server can be correctly parsed by the client.

Close #22