Fix initialization regression caused by #30
When using the streamable HTTP transport, the client should not block on opening the SSE stream as this is optional. We still need to block in the legacy HTTP+SSE case because the server needs to send the client the `endpoint` event over SSE before the client is initialized.

Close #37